+++ src/write.ts
